SCR: Modify long type in int

+++ src/utils/Pow.java
@@ -2,8 +2,8 @@ package utils;
 
 public class Pow {
 
-    public static long pow(int a, int b) {
-        long finalProduct = 1;
+    public static int pow(int a, int b) {
+        int finalProduct = 1;
         int more = b;
         while (more != 0) {
             finalProduct *= a;
